VE cell interpolation!
The program is now significantly more accurate to the real-world use cases of the VE table. The PCM applies interpolation when it is between a VE cell, and this new version undoes that interpolation in the log file to produce a scarily accurate result. You should see a marked increase in accuracy of observed AFR, as well as corrections.